+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0 (wip)
 ## API breaks
 * `Console.print_color` : replace %{} color marker with #[]
+* `resize` function added to `Engine` trait to allow console resize
 
 # 0.1.0 - 2018 Aug 3
 First release